Token update system changed. Delay for requests reduced to 300L.

+++ app/src/main/java/com/ds/money_manager/data/repository/handler/MoneyManagerDataHandlerImpl.kt
@@ -1,71 +1,68 @@
 package com.ds.money_manager.data.repository.handler
 
+import com.ds.money_manager.Constants.UNAUTHORIZED_CODE
 import com.ds.money_manager.data.model.api.*
 import com.ds.money_manager.data.repository.api.MoneyManagerApi
+import com.ds.money_manager.data.repository.preferences.AppSharedPreferences
 import com.ds.money_manager.utils.ApiErrorUtils
 import com.ds.money_manager.utils.ApiException
+import okhttp3.Response
+import okhttp3.ResponseBody
 import java.math.BigDecimal
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
 import javax.inject.Inject
 
-class MoneyManagerDataHandlerImpl @Inject constructor(val moneyManagerApi: MoneyManagerApi) :
-    MoneyManagerDataHandler {
+class MoneyManagerDataHandlerImpl @Inject constructor(
+    private val moneyManagerApi: MoneyManagerApi,
+    private val preferences: AppSharedPreferences
+) : MoneyManagerDataHandler {
 
     private val UNKNOWN_ERROR = "Unknown error"
 
     override fun signIn(name: String, password: String): SignInResponse {
         val result = moneyManagerApi.signIn(SignInRequest(name, password)).execute()
+
         if (result.isSuccessful && result.body() != null)
             return result.body()!!
         else {
-            val apiError = ApiErrorUtils.parseError(result.errorBody()!!)
-            apiError?.let { throw ApiException(it.title, it.description, result.code()) }
-                ?: throw Exception(UNKNOWN_ERROR)
+            throw configureError(result.errorBody()!!, result.code())
         }
     }
 
     override fun signUp(name: String, password: String): SignUpResponse {
         val result = moneyManagerApi.signUp(SignUpRequest(name, password)).execute()
-        if (result.isSuccessful && result.body() != null)
-            return result.body()!!
-        else {
-            val apiError = ApiErrorUtils.parseError(result.errorBody()!!)
-            apiError?.let { throw ApiException(it.title, it.description, result.code()) }
-                ?: throw Exception(UNKNOWN_ERROR)
-        }
-    }
 
-    override fun checkToken(): Boolean {
-        val result = moneyManagerApi.checkToken().execute()
         if (result.isSuccessful && result.body() != null)
             return result.body()!!
         else {
-            val apiError = ApiErrorUtils.parseError(result.errorBody()!!)
-            apiError?.let { throw ApiException(it.title, it.description, result.code()) }
-                ?: throw Exception(UNKNOWN_ERROR)
+            throw configureError(result.errorBody()!!, result.code())
         }
     }
 
     override fun getTotalBalance(): BigDecimal {
-        val result = moneyManagerApi.getTotalBalance().execute()
+        var result = moneyManagerApi.getTotalBalance().execute()
+        if (result.code() == UNAUTHORIZED_CODE) {
+            updateToken()
+            result = moneyManagerApi.getTotalBalance().execute()
+        }
         if (result.isSuccessful && result.body() != null)
             return result.body()!!.totalBalance
         else {
-            val apiError = ApiErrorUtils.parseError(result.errorBody()!!)
-            apiError?.let { throw ApiException(it.title, it.description, result.code()) }
-                ?: throw Exception(UNKNOWN_ERROR)
+            throw configureError(result.errorBody()!!, result.code())
         }
     }
 
     override fun getAllWalletsDetails(): List<WalletDetailsResponse> {
-        val result = moneyManagerApi.getAllWalletsDetails().execute()
+        var result = moneyManagerApi.getAllWalletsDetails().execute()
+        if (result.code() == UNAUTHORIZED_CODE) {
+            updateToken()
+            result = moneyManagerApi.getAllWalletsDetails().execute()
+        }
         if (result.isSuccessful && result.body() != null)
             return result.body()!!
         else {
-            val apiError = ApiErrorUtils.parseError(result.errorBody()!!)
-            apiError?.let { throw ApiException(it.title, it.description, result.code()) }
-                ?: throw Exception(UNKNOWN_ERROR)
+            throw configureError(result.errorBody()!!, result.code())
         }
     }
 
@@ -75,16 +72,22 @@ class MoneyManagerDataHandlerImpl @Inject constructor(val moneyManagerApi: Money
     ): List<StatisticItemResponse> {
         val dateTimeFormatter = DateTimeFormatter.ofPattern("yyyy-MM-dd")
 
-        val result = moneyManagerApi.getTotalStatisticByDate(
+        var result = moneyManagerApi.getTotalStatisticByDate(
             dateFrom.format(dateTimeFormatter),
             dateTo.format(dateTimeFormatter)
         ).execute()
+
+        if (result.code() == UNAUTHORIZED_CODE) {
+            updateToken()
+            result = moneyManagerApi.getTotalStatisticByDate(
+                dateFrom.format(dateTimeFormatter),
+                dateTo.format(dateTimeFormatter)
+            ).execute()
+        }
         if (result.isSuccessful && result.body() != null)
             return result.body()!!
         else {
-            val apiError = ApiErrorUtils.parseError(result.errorBody()!!)
-            apiError?.let { throw ApiException(it.title, it.description, result.code()) }
-                ?: throw Exception(UNKNOWN_ERROR)
+            throw configureError(result.errorBody()!!, result.code())
         }
     }
 
@@ -95,69 +98,90 @@ class MoneyManagerDataHandlerImpl @Inject constructor(val moneyManagerApi: Money
         offset: Int,
         limit: Int
     ): List<TransactionResponse> {
-        val result =
-            moneyManagerApi.getAllTransactionsByDate(dateFrom, dateTo, walletId, offset, limit).execute()
+        var result =
+            moneyManagerApi.getAllTransactionsByDate(dateFrom, dateTo, walletId, offset, limit)
+                .execute()
+
+        if (result.code() == UNAUTHORIZED_CODE) {
+            updateToken()
+            result =
+                moneyManagerApi.getAllTransactionsByDate(dateFrom, dateTo, walletId, offset, limit)
+                    .execute()
+        }
         if (result.isSuccessful && result.body() != null)
             return result.body()!!
         else {
-            val apiError = ApiErrorUtils.parseError(result.errorBody()!!)
-            apiError?.let { throw ApiException(it.title, it.description, result.code()) }
-                ?: throw Exception(UNKNOWN_ERROR)
+            throw configureError(result.errorBody()!!, result.code())
         }
     }
 
     override fun getLastTransactions(limit: Int): List<TransactionResponse> {
-        val result = moneyManagerApi.getLastTransactions(limit).execute()
+        var result = moneyManagerApi.getLastTransactions(limit).execute()
+
+        if (result.code() == UNAUTHORIZED_CODE) {
+            updateToken()
+            result = moneyManagerApi.getLastTransactions(limit).execute()
+        }
         if (result.isSuccessful && result.body() != null)
             return result.body()!!
         else {
-            val apiError = ApiErrorUtils.parseError(result.errorBody()!!)
-            apiError?.let { throw ApiException(it.title, it.description, result.code()) }
-                ?: throw Exception(UNKNOWN_ERROR)
+            throw configureError(result.errorBody()!!, result.code())
         }
     }
 
     override fun postWallet(name: String, currencyId: Int): WalletResponse {
-        val result = moneyManagerApi.postWallet(WalletRequest(name, currencyId)).execute()
+        var result = moneyManagerApi.postWallet(WalletRequest(name, currencyId)).execute()
+
+        if (result.code() == UNAUTHORIZED_CODE) {
+            updateToken()
+            result = moneyManagerApi.postWallet(WalletRequest(name, currencyId)).execute()
+        }
         if (result.isSuccessful && result.body() != null)
             return result.body()!!
         else {
-            val apiError = ApiErrorUtils.parseError(result.errorBody()!!)
-            apiError?.let { throw ApiException(it.title, it.description, result.code()) }
-                ?: throw Exception(UNKNOWN_ERROR)
+            throw configureError(result.errorBody()!!, result.code())
         }
     }
 
     override fun putWallet(walletId: Int, name: String): WalletUpdateResponse {
-        val result = moneyManagerApi.putWallet(WalletUpdateRequest(walletId, name)).execute()
+        var result = moneyManagerApi.putWallet(WalletUpdateRequest(walletId, name)).execute()
+
+        if (result.code() == UNAUTHORIZED_CODE) {
+            updateToken()
+            result = moneyManagerApi.putWallet(WalletUpdateRequest(walletId, name)).execute()
+        }
         if (result.isSuccessful && result.body() != null)
             return result.body()!!
         else {
-            val apiError = ApiErrorUtils.parseError(result.errorBody()!!)
-            apiError?.let { throw ApiException(it.title, it.description, result.code()) }
-                ?: throw Exception(UNKNOWN_ERROR)
+            throw configureError(result.errorBody()!!, result.code())
         }
     }
 
     override fun deleteWallet(walletId: Int) {
-        val result = moneyManagerApi.deleteWallet(walletId).execute()
+        var result = moneyManagerApi.deleteWallet(walletId).execute()
+
+        if (result.code() == UNAUTHORIZED_CODE) {
+            updateToken()
+            result = moneyManagerApi.deleteWallet(walletId).execute()
+        }
         if (result.isSuccessful)
             return
         else {
-            val apiError = ApiErrorUtils.parseError(result.errorBody()!!)
-            apiError?.let { throw ApiException(it.title, it.description, result.code()) }
-                ?: throw Exception(UNKNOWN_ERROR)
+            throw configureError(result.errorBody()!!, result.code())
         }
     }
 
     override fun getWallet(walletId: Int): WalletResponse {
-        val result = moneyManagerApi.getWallet(walletId).execute()
+        var result = moneyManagerApi.getWallet(walletId).execute()
+
+        if (result.code() == UNAUTHORIZED_CODE) {
+            updateToken()
+            result = moneyManagerApi.getWallet(walletId).execute()
+        }
         if (result.isSuccessful && result.body() != null)
             return result.body()!!
         else {
-            val apiError = ApiErrorUtils.parseError(result.errorBody()!!)
-            apiError?.let { throw ApiException(it.title, it.description, result.code()) }
-                ?: throw Exception(UNKNOWN_ERROR)
+            throw configureError(result.errorBody()!!, result.code())
         }
     }
 
@@ -167,13 +191,16 @@ class MoneyManagerDataHandlerImpl @Inject constructor(val moneyManagerApi: Money
         sum: BigDecimal,
         date: String
     ): IncomeResponse {
-        val result = moneyManagerApi.postIncome(IncomeRequest(walletId, name, sum, date)).execute()
+        var result = moneyManagerApi.postIncome(IncomeRequest(walletId, name, sum, date)).execute()
+
+        if (result.code() == UNAUTHORIZED_CODE) {
+            updateToken()
+            result = moneyManagerApi.postIncome(IncomeRequest(walletId, name, sum, date)).execute()
+        }
         if (result.isSuccessful && result.body() != null)
             return result.body()!!
         else {
-            val apiError = ApiErrorUtils.parseError(result.errorBody()!!)
-            apiError?.let { throw ApiException(it.title, it.description, result.code()) }
-                ?: throw Exception(UNKNOWN_ERROR)
+            throw configureError(result.errorBody()!!, result.code())
         }
     }
 
@@ -184,37 +211,48 @@ class MoneyManagerDataHandlerImpl @Inject constructor(val moneyManagerApi: Money
         sum: BigDecimal,
         date: String
     ): IncomeResponse {
-        val result =
+        var result =
             moneyManagerApi.putIncome(IncomeUpdateRequest(incomeId, walletId, name, sum, date))
                 .execute()
+
+        if (result.code() == UNAUTHORIZED_CODE) {
+            updateToken()
+            result =
+                moneyManagerApi.putIncome(IncomeUpdateRequest(incomeId, walletId, name, sum, date))
+                    .execute()
+        }
         if (result.isSuccessful && result.body() != null)
             return result.body()!!
         else {
-            val apiError = ApiErrorUtils.parseError(result.errorBody()!!)
-            apiError?.let { throw ApiException(it.title, it.description, result.code()) }
-                ?: throw Exception(UNKNOWN_ERROR)
+            throw configureError(result.errorBody()!!, result.code())
         }
     }
 
     override fun deleteIncome(incomeId: Int) {
-        val result = moneyManagerApi.deleteIncome(incomeId).execute()
+        var result = moneyManagerApi.deleteIncome(incomeId).execute()
+
+        if (result.code() == UNAUTHORIZED_CODE) {
+            updateToken()
+            result = moneyManagerApi.deleteIncome(incomeId).execute()
+        }
         if (result.isSuccessful)
             return
         else {
-            val apiError = ApiErrorUtils.parseError(result.errorBody()!!)
-            apiError?.let { throw ApiException(it.title, it.description, result.code()) }
-                ?: throw Exception(UNKNOWN_ERROR)
+            throw configureError(result.errorBody()!!, result.code())
         }
     }
 
     override fun getCostTypes(): List<CostTypeResponse> {
-        val result = moneyManagerApi.getCostTypes().execute()
+        var result = moneyManagerApi.getCostTypes().execute()
+
+        if (result.code() == UNAUTHORIZED_CODE) {
+            updateToken()
+            result = moneyManagerApi.getCostTypes().execute()
+        }
         if (result.isSuccessful && result.body() != null)
             return result.body()!!
         else {
-            val apiError = ApiErrorUtils.parseError(result.errorBody()!!)
-            apiError?.let { throw ApiException(it.title, it.description, result.code()) }
-                ?: throw Exception(UNKNOWN_ERROR)
+            throw configureError(result.errorBody()!!, result.code())
         }
     }
 
@@ -225,14 +263,19 @@ class MoneyManagerDataHandlerImpl @Inject constructor(val moneyManagerApi: Money
         walletId: Int,
         costTypeId: Int
     ): CostResponse {
-        val result =
+        var result =
             moneyManagerApi.postCost(CostRequest(name, sum, date, walletId, costTypeId)).execute()
+
+        if (result.code() == UNAUTHORIZED_CODE) {
+            updateToken()
+            result =
+                moneyManagerApi.postCost(CostRequest(name, sum, date, walletId, costTypeId))
+                    .execute()
+        }
         if (result.isSuccessful && result.body() != null)
             return result.body()!!
         else {
-            val apiError = ApiErrorUtils.parseError(result.errorBody()!!)
-            apiError?.let { throw ApiException(it.title, it.description, result.code()) }
-                ?: throw Exception(UNKNOWN_ERROR)
+            throw configureError(result.errorBody()!!, result.code())
         }
     }
 
@@ -244,7 +287,7 @@ class MoneyManagerDataHandlerImpl @Inject constructor(val moneyManagerApi: Money
         walledId: Int,
         costTypeId: Int
     ): CostResponse {
-        val result =
+        var result =
             moneyManagerApi.putCost(
                 CostUpdateRequest(
                     costId,
@@ -254,47 +297,81 @@ class MoneyManagerDataHandlerImpl @Inject constructor(val moneyManagerApi: Money
                     walledId,
                     costTypeId
                 )
-            )
-                .execute()
+            ).execute()
+
+        if (result.code() == UNAUTHORIZED_CODE) {
+            updateToken()
+            result =
+                moneyManagerApi.putCost(
+                    CostUpdateRequest(
+                        costId,
+                        name,
+                        sum,
+                        date,
+                        walledId,
+                        costTypeId
+                    )
+                ).execute()
+        }
         if (result.isSuccessful && result.body() != null)
             return result.body()!!
         else {
-            val apiError = ApiErrorUtils.parseError(result.errorBody()!!)
-            apiError?.let { throw ApiException(it.title, it.description, result.code()) }
-                ?: throw Exception(UNKNOWN_ERROR)
+            throw configureError(result.errorBody()!!, result.code())
         }
     }
 
     override fun deleteCost(costId: Int) {
-        val result = moneyManagerApi.deleteCost(costId).execute()
+        var result = moneyManagerApi.deleteCost(costId).execute()
+
+        if (result.code() == UNAUTHORIZED_CODE) {
+            updateToken()
+            result = moneyManagerApi.deleteCost(costId).execute()
+        }
         if (result.isSuccessful)
             return
         else {
-            val apiError = ApiErrorUtils.parseError(result.errorBody()!!)
-            apiError?.let { throw ApiException(it.title, it.description, result.code()) }
-                ?: throw Exception(UNKNOWN_ERROR)
+            throw configureError(result.errorBody()!!, result.code())
         }
     }
 
     override fun getCost(costId: Int): CostResponse {
-        val result = moneyManagerApi.getCost(costId).execute()
+        var result = moneyManagerApi.getCost(costId).execute()
+
+        if (result.code() == UNAUTHORIZED_CODE) {
+            updateToken()
+            result = moneyManagerApi.getCost(costId).execute()
+        }
         if (result.isSuccessful && result.body() != null)
             return result.body()!!
         else {
-            val apiError = ApiErrorUtils.parseError(result.errorBody()!!)
-            apiError?.let { throw ApiException(it.title, it.description, result.code()) }
-                ?: throw Exception(UNKNOWN_ERROR)
+            throw configureError(result.errorBody()!!, result.code())
         }
     }
 
     override fun getIncome(incomeId: Int): IncomeResponse {
-        val result = moneyManagerApi.getIncome(incomeId).execute()
+        var result = moneyManagerApi.getIncome(incomeId).execute()
+
+        if (result.code() == UNAUTHORIZED_CODE) {
+            updateToken()
+            result = moneyManagerApi.getIncome(incomeId).execute()
+        }
         if (result.isSuccessful && result.body() != null)
             return result.body()!!
         else {
-            val apiError = ApiErrorUtils.parseError(result.errorBody()!!)
-            apiError?.let { throw ApiException(it.title, it.description, result.code()) }
-                ?: throw Exception(UNKNOWN_ERROR)
+            throw configureError(result.errorBody()!!, result.code())
         }
+    }
+
+    private fun configureError(errorBody: ResponseBody, code: Int): Exception {
+        val apiError = ApiErrorUtils.parseError(errorBody)
+        return if (apiError != null)
+            ApiException(apiError.title, apiError.description, code)
+        else
+            Exception(UNKNOWN_ERROR)
+    }
+
+    private fun updateToken() {
+        val response = signIn(preferences.username, preferences.password)
+        preferences.token = response.token
     }
 }
